Remove derive-try-from-primitive, reorganise types
