+++ flash-lso/src/amf0/type_marker.rs
@@ -1,8 +1,6 @@
-use derive_try_from_primitive::TryFromPrimitive;
-
 /// Type markers used in AMF0
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(TryFromPrimitive, Eq, PartialEq, Debug, Copy, Clone)]
+#[derive(Eq, PartialEq, Debug, Copy, Clone)]
 #[repr(u8)]
 pub(crate) enum TypeMarker {
     /// Number
@@ -41,4 +39,32 @@ pub(crate) enum TypeMarker {
     TypedObject = 16,
     /// Embedded AMF3 element
     AMF3 = 17,
+}
+
+impl TryFrom<u8> for TypeMarker {
+    type Error = ();
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0 => Ok(Self::Number),
+            1 => Ok(Self::Boolean),
+            2 => Ok(Self::String),
+            3 => Ok(Self::Object),
+            4 => Ok(Self::MovieClip),
+            5 => Ok(Self::Null),
+            6 => Ok(Self::Undefined),
+            7 => Ok(Self::Reference),
+            8 => Ok(Self::MixedArrayStart),
+            9 => Ok(Self::ObjectEnd),
+            10 => Ok(Self::Array),
+            11 => Ok(Self::Date),
+            12 => Ok(Self::LongString),
+            13 => Ok(Self::Unsupported),
+            14 => Ok(Self::RecordSet),
+            15 => Ok(Self::Xml),
+            16 => Ok(Self::TypedObject),
+            17 => Ok(Self::AMF3),
+            _ => Err(())
+        }
+    }
 }
