+++ flash-lso/src/amf3/type_marker.rs
@@ -1,8 +1,6 @@
-use derive_try_from_primitive::TryFromPrimitive;
-
 /// Type markers used in AMF3
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(TryFromPrimitive, Eq, PartialEq, Debug, Copy, Clone)]
+#[derive(Eq, PartialEq, Debug, Copy, Clone)]
 #[repr(u8)]
 pub(crate) enum TypeMarker {
     /// Undefined
@@ -41,4 +39,32 @@ pub(crate) enum TypeMarker {
     VectorObject = 0x10,
     /// Dictionary
     Dictionary = 0x11,
+}
+
+impl TryFrom<u8> for TypeMarker {
+    type Error = ();
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0x00 => Ok(Self::Undefined),
+            0x01 => Ok(Self::Null),
+            0x02 => Ok(Self::False),
+            0x03 => Ok(Self::True),
+            0x04 => Ok(Self::Integer),
+            0x05 => Ok(Self::Number),
+            0x06 => Ok(Self::String),
+            0x07 => Ok(Self::Xml),
+            0x08 => Ok(Self::Date),
+            0x09 => Ok(Self::Array),
+            0x0A => Ok(Self::Object),
+            0x0B => Ok(Self::XmlString),
+            0x0C => Ok(Self::ByteArray),
+            0x0D => Ok(Self::VectorInt),
+            0x0E => Ok(Self::VectorUInt),
+            0x0F => Ok(Self::VectorDouble),
+            0x10 => Ok(Self::VectorObject),
+            0x11 => Ok(Self::Dictionary),
+            _ => Err(())
+        }
+    }
 }
